refactor: simplify StorageManager naming
