+++ transfer_queue/metadata.py
@@ -30,11 +30,6 @@ from transfer_queue.utils.logging_utils import get_logger
 logger = get_logger(__name__)
 
 
-# ---------------------------------------------------------------------------
-# Internal helpers
-# ---------------------------------------------------------------------------
-
-
 def _extra_info_values_equal(a: Any, b: Any) -> bool:
     """Compare two extra_info values for equality.
 
@@ -55,7 +50,7 @@ def _extra_info_values_equal(a: Any, b: Any) -> bool:
 class _SampleView:
     """Lazy read-only view of a single sample row in a columnar BatchMeta.
 
-    All returned dicts are ``MappingProxyType`` – attempts to mutate them
+    All returned dicts are ``MappingProxyType``, and attempts to mutate them
     raise ``TypeError``, making it obvious that this is a snapshot view.
     """
 
@@ -379,7 +374,6 @@ class BatchMeta:
         return [meta.get("shape")] * self.size
 
     # ==================== Extra Info Methods ====================
-
     def get_extra_info(self, key: str, default: Any = None) -> Any:
         """Get extra info by key"""
         return self.extra_info.get(key, default)
@@ -417,7 +411,6 @@ class BatchMeta:
         return key in self.extra_info
 
     # ==================== Custom Meta Methods (User Layer) ====================
-
     def get_all_custom_meta(self) -> list[dict[str, Any]]:
         """Get all custom_meta as a list of dictionary (one per sample, in global_indexes order).
 
@@ -451,7 +444,6 @@ class BatchMeta:
         self.custom_meta = [{} for _ in range(self.size)]
 
     # ==================== Core BatchMeta Operations ====================
-
     def add_fields(self, tensor_dict: TensorDict, set_all_ready: bool = True) -> "BatchMeta":
         """Add new fields from a TensorDict to all samples in this batch.
         This modifies the batch in-place to include the new fields.
