+++ transfer_queue/storage/clients/__init__.py
@@ -14,14 +14,13 @@
 # limitations under the License.
 
 # This module is currently empty but reserved for future client implementations
-from .base import TransferQueueStorageKVClient
-from .factory import StorageClientFactory
+from .base import StorageClientFactory, StorageKVClient
 from .mooncake_client import MooncakeStoreClient
 from .ray_storage_client import RayStorageClient
 from .yuanrong_client import YuanrongStorageClient
 
 __all__ = [
-    "TransferQueueStorageKVClient",
+    "StorageKVClient",
     "StorageClientFactory",
     "RayStorageClient",
     "MooncakeStoreClient",
